refactor: move logic from SavedCartsListModal into SavedCartsTable component
