+++ messages/es.json
@@ -47,7 +47,8 @@
   "store/checkout.b2b.column.refId": "#",
   "store/checkout.b2b.column.seller": "Vendedor",
   "store/checkout.b2b.column.margin": "Margen",
-  "store/checkout.b2b.column.createdIn": "creado en",
+  "store/checkout.b2b.column.createdIn": "Creado en",
+  "store/checkout.b2b.column.item": "Elementos",
 
   "store/checkout.b2b.button.backToHome": "Volver a la página de inicio",
   "store/checkout.b2b.button.clearCart": "Limpiar carrito",
